% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binsurvs}
\alias{binsurvs}
\title{Simulated Binomial Surveys Data}
\format{
`binsurvs`
A list with two elements:
\describe{
  \item{X}{The surveys with three items. `item 1` has no DIF effects; `item 2` has uniform DIF effects; `item 3` has non-uniform DIF effects}
  \item{Z}{Two normalized exogenous covariates: age and sex}
}
}
\usage{
binsurvs
}
\description{
The simulated data used for examples.
}
\keyword{datasets}
