% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICC.R
\name{getICC}
\alias{getICC}
\title{Plotting the Item Characteristics Curve (ICC) based on LIDIF results}
\usage{
getICC(
  coefs_list,
  cov_mat,
  compare_var,
  type_list,
  title = NULL,
  latent_name = "latent",
  xrange = seq(-3, 3, 0.01)
)
}
\arguments{
\item{coefs_list}{The list of DIF effects. Should be of the same format as the `coefficients` value in `LIDIF` output.}

\item{cov_mat}{The covariate matrix. DIF effects can be visualized by specifying different values for one of the covariates.}

\item{compare_var}{The name of covariate you would to visualize the DIF effect.}

\item{type_list}{A vector of characters containing the types of items (`binary`,`ordinal`). If the length of vector is 1, then the same type would apply to all items.}

\item{title}{Title of the plot. The defualt value is `NULL`.}

\item{latent_name}{Name of the latent variable (showing in the x-axises). The default is `latent`.}

\item{xrange}{the seqeunce of the latent variable to be plotted from. The default setting is from -3 to 3 and seperated by 0.01.}
}
\value{
The plot combining ICC curves of all items.
}
\description{
The function plots the ICC curve by items and is especially helpful if you want to visualize uniform/non-uniform DIF effects found in LIDIF modeling.
}
\examples{
 \dontrun{
## load the sample data (see `?LIDIF::binsurvs` for details)
binsurvs = LIDIF::binsurvs # The exogenous variables are age and sex.


## prepare for model fitting
dat.list = prepare_data(binsurvs$X, binsurvs$Z)

## LIDIF fitting
# The parameter specified is set to reduce the running time.
# In your analysis, please refer to the default value.
library(LIDIF)
res = LIDIF(dat.list = dat.list,
            cl_num = 5, # before running, please check if you have 5 cores
            type_list = "binary",
            maxit = 2,
            random_per = 0.5,
            init_nums = 2,
            init_maxit = 2)

## Plot the DIF effects for sex using ICC curves
# specify the covariate matrix
tt = cbind("age" = c(mean(dat.list$Z[,"age"]), mean(dat.list$Z[,"age"])),
           "sex" = c(0,1))
getICC(res$coefficients,
      cov_mat = tt,
      compare_var = "sex",
      type_list = "binary")
}

}
