% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIDIF.R
\name{LIDIF}
\alias{LIDIF}
\title{Performing Likelihood-based Investigation of Differential Item Functioning  (LIDIF)}
\usage{
LIDIF(
  dat.list,
  cl_num,
  type_list,
  output_dir = NULL,
  int_bound = 3L,
  evalimit = 10L,
  stop_at = 0.01,
  maxit = 50L,
  optim_trace = TRUE,
  optim_reltol = 0.1,
  hessian = FALSE,
  variance = TRUE,
  initialization = TRUE,
  init_input = NULL,
  random_per = 0.05,
  replace = FALSE,
  init_nums = 30,
  init_maxit = 5
)
}
\arguments{
\item{dat.list}{The data list containing both the observed items matrix and enogenous variable matrix derived from `prepare_data` function.}

\item{cl_num}{Number of cores LIDIF can use. The more the better!}

\item{type_list}{A vector of characters containing the types of items (`binary`,`ordinal`). If the length of vector is 1, then the same type would apply to all items.}

\item{output_dir}{The path to save intermediate results. The default value is NULL. If NULL, the results will be save in current working directory.}

\item{int_bound}{The adaptive integration boundaries. The default value is 3, meaning the integration taking from -3 to 3.}

\item{evalimit}{The number of quadrature used in the adaptive integration. The default value is 20.}

\item{stop_at}{The absolute convergence tolerance. The default value is 0.01.}

\item{maxit}{The maximum number of iterations used in the EM algorithm. The default value is 50.}

\item{optim_trace}{Logical. Should track optimization process? The default value is TRUE.}

\item{optim_reltol}{The relative convergence tolerance for M-step. The default value is 0.1.}

\item{hessian}{Logical. Should return the numerical hessian matrix in optimization procedures? The default value is FALSE.}

\item{variance}{Logical. Should calculate variance and covariance matrix in the end? The default value is TRUE.}

\item{initialization}{Logical. Should start from random initialization? The default is TRUE.}

\item{init_input}{If initialization = FALSE, specify your own initialization here. The default is NULL.}

\item{random_per}{Percentage of random samples in initialization. The default value is 0.05.}

\item{replace}{Logical. Should generate random samples with replacement? The default is FALSE.}

\item{init_nums}{Number of random initialization. The default is 30.}

\item{init_maxit}{Maximum number of iterations in initialization. The default is 5.}
}
\value{
An object of the class `LIDIF` that contains
\itemize{
  \item `coefficients`: a list of the estimated DIF effects. with each sublist corresponding to an item.
  \item `variance`: A list of the estimated variance of the individual coefficients, following the same order as `coefficients`. Return only if variance = TRUE.
  \item `cov`: The estimated variance-covariance matrix for all the coefficeints. Return only if variance = TRUE.

}
}
\description{
The function takes the output of `prepare_data` function and run the LIDIF procedure. Notice that the procedure incoporats parallel computing functions to boost the computation speed. Therefore, we strongly recommend run the function with multiple cores. Please refer to the arguments on specifying number of cores.
}
\details{
Denote the latent variable as \eqn{Y}, the covariate matrix as \eqn{Z} and \eqn{X_1,...,X_p} are the observed items.
For items with binary responses (0 or 1), LIDIF models
\deqn{logit(P(X_i =1)) = \alpha_i+\beta_i^TZ + \zeta_i^TZ \times Y}
For items with ordinal responses (1, 2,...,\eqn{M_i}), LIDIF models
\deqn{logit(P(X_i \geq m)) = \alpha_{im}+\beta_i^TZ + \zeta_i^TZ \times Y}
where \eqn{m \in \{1, 2,...,M_i-1\}} and \eqn{\alpha_{i1} > ...> \alpha_{im-1}}.\cr

The model returns maximum likelihood estimators of all the coefficients and their estimated variance (by choice).\cr

The key to the estimation accuracy of LIDIF is specifying multiple random initialization in `init_nums` arguments. A trade-off is the more
number of random initialization you specified, the longer it took to run `LIDIF`. But you could always speed up the process by running on multiple
computing cores (specified in `cl_num`) arguments.\cr

For your reference, the default setting provides accurate estimation in simulations with 5 items and 2 exogenous covariates from 1000 observations.
And the computing time was 1 hour with 5 cores using R version 4.0.2.
}
\examples{
 \dontrun{
## load the sample data (see `?LIDIF::binsurvs` for details)
binsurvs = LIDIF::binsurvs # The exogenous variables are age and sex.

## prepare for model fitting
dat.list = prepare_data(binsurvs$X, binsurvs$Z)

## LIDIF fitting
# The parameter specified is set to reduce the running time.
# In your analysis, please refer to the default value.
library(LIDIF)
res = LIDIF(dat.list = dat.list,
            cl_num = 5, # before running, please check if you have 5 cores
            type_list = "binary",
            maxit = 2,
            random_per = 0.5,
            init_nums = 2,
            init_maxit = 2)
}
}
