% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Preparing data for LIDIF function.}
\usage{
prepare_data(X, Z)
}
\arguments{
\item{X}{A numeric matrix with columns corresponding to items. If the item is binary, please code it as 0/1; if the item is ordinal, please code it as integers starting from 1, e.g. 1,2,3. X can be missing at some columns but not the entire row.}

\item{Z}{the matrix of exogenous variables. Z should not contain any missing data and should have the same number of rows as X. Please do not include intercept here.}
}
\value{
A list containing two elements.
}
\description{
The function prepare your raw data to fit into LIDIF model.
}
\examples{
items_matrix = cbind.data.frame("item1" = rbinom(100, 0,1),
                                "item2" = sample(c(1,2,3), size = 100, replace = TRUE))
covariates = cbind.data.frame("age" = runif(100, 10, 50),
                             "sex" = rbinom(100, 0, 1))
prepare_data(X = items_matrix, Z = covariates)


}
